refactor: move main page to be slug page
